+++ umd/recoil.js
@@ -3794,6 +3794,11 @@ This is currently a DEV-only warning but will become a thrown exception in the n
     getID() {
       this.checkRefCount_INTERNAL();
       return this._store.getState().currentTree.stateID;
+    }
+
+    getStoreID() {
+      this.checkRefCount_INTERNAL();
+      return this._store.storeID;
     } // We want to allow the methods to be destructured and used as accessors
     // eslint-disable-next-line fb-www/extra-arrow-initializer